+++ addons/base/models/__init__.py
@@ -2,3 +2,6 @@ from . import res_users
 from . import res_partner
 from . import res_users_apikeys
 from . import ir_cron
+from . import ir_sequence
+from . import ir_attachment
+from . import ir_model